Doxygen styled comments in OpenGLTools headers
Added Doxygen styled comments in OpenGLTools library header files to allow generating documentation using Doxygen

+++ src/UniformBufferObject.cpp
@@ -38,17 +38,17 @@ namespace OGL {
     }
 
     void UniformBufferObject::setBindingPoint(
-        size_t pointIndex
+        GLuint pointIndex
     ) {
-        glBindBufferBase(GL_UNIFORM_BUFFER, static_cast<GLuint>(pointIndex), m_descriptor);
+        glBindBufferBase(GL_UNIFORM_BUFFER, pointIndex, m_descriptor);
     }
 
     void UniformBufferObject::setBindingPoint( 
-        size_t pointIndex, 
+        GLuint pointIndex,
         size_t offset, 
         size_t size
     ) {
-        glBindBufferRange(GL_UNIFORM_BUFFER, static_cast<GLuint>(pointIndex), m_descriptor, offset, size);
+        glBindBufferRange(GL_UNIFORM_BUFFER, pointIndex, m_descriptor, offset, size);
     }
 
 } // OGL
